requirements-dev.txt
-r requirements.txt
pytest==8.3.4
python-dotenv==1.0.1
black==25.1.0
twine==6.1.0
